perf(dsv4): replace torch RMSNorm with C++ rtp_llm_ops.rmsnorm (bf16 weight)
Every ``_RMSNorm`` site in dsv4 was a 6-launch torch chain:
``.float().square().mean(-1).rsqrt().mul().to()``.  Route through the
framework C++ op instead (single launch).  Covers:

- ``block.py`` _RMSNorm (attn_norm / ffn_norm; MTPBlock enorm / hnorm / norm)
- ``transformer.py`` _RMSNorm (final norm before lm_head)
- ``compressor.py`` _CompressorNorm (``_rmsnorm`` method)

Weight dtype switches fp32 -> bf16 to match vLLM / SGLang inference
(vLLM ``layernorm.py`` default is ``torch.get_default_dtype()`` = bf16;
confirmed via vLLM source).  Checkpoint is already bf16; DeepSeek's
reference inference/model.py kept fp32 per the author's own "for
convenient" comment.  The C++ ``rtp_llm_ops.rmsnorm`` requires bf16
weight (silent NaN on fp32).  Weight loaders in block.py / compressor.py
adjusted from ``.float()`` to ``.to(torch.bfloat16)`` accordingly.

Perf (GPU 7 / B300, dim=4096, live module incl. reshape overhead):

  T=1     19.9us -> 16.4us  (1.22x)
  T=8     20.6us -> 15.4us  (1.34x)
  T=128   19.0us -> 14.9us  (1.28x)
  T=1024  20.0us -> 14.8us  (1.35x)
  T=4096  19.2us -> 16.5us  (1.17x)

UT ``test_rmsnorm_replace.py`` (new): diff <=1 bf16 ULP, mean <=2e-8;
C++ 2.7-12x vs torch across decode/prefill shapes.

+++ rtp_llm/models_py/modules/dsv4/transformer.py
@@ -220,7 +220,7 @@ class V4Transformer(nn.Module):
                 weights["head.weight"].float(), requires_grad=False
             )
             self.norm.weight = nn.Parameter(
-                weights["norm.weight"].float(), requires_grad=False
+                weights["norm.weight"].to(torch.bfloat16), requires_grad=False
             )
             self.hc_head_fn = nn.Parameter(
                 weights["hc_head_fn"].float(), requires_grad=False
